Moved "timed()" to own file.
